Add license to bundle output
Thanks for the tip, random github user!

(I briefly wondered whether the lack of a license marker on this file is
any different from the millions of other files that can be downloaded
trivially from GitHub without their license. But  rather than ponder,
it's more helpful to actually add a bit of content which both puts the
license in place, and points out what on earth this file is and where it
came from - something that I'm sure more than one inheritor of a project
has been mystified by.)